refactor io module
