+++ tests/io/test_molfile_reader.py
@@ -1,12 +1,10 @@
 import pytest
-
-from tucan.io import (
+from tucan.io import graph_from_molfile_text, MolfileParserException
+from tucan.io.molfile_reader import (
+    _concat_lines_with_dash,
     _parse_atom_block_molfile3000,
     _parse_bond_block_molfile3000,
     _read_file,
-    MolfileParserException,
-    graph_from_molfile_text,
-    _concat_lines_with_dash,
 )
